Extract graph orchestration from __init__ into _observe
__init__ now just re-exports the public surface (observe_node,
NodeNotFoundError, latched_qos); the graph polling and endpoint collection
move to _observe, matching the _-prefixed private-module pattern already
used for _endpoints/_parameters/_qos.  No behaviour change.
